commit 132f92b0e6c7ef46ac10daee56575ea421313e21
    12596 A typo in usr/src/man/man1/ls.1
commit c3267a7bdebfa73aeda8a9cc71aa02f5420ba00c
    12594 bge device IDs do not have a leading 0x
commit 7112648bd7986a88f0ffa61263d5cbf7229d8b03
    12472 pam_list does not have 'group' option
commit fba27d8741c08c38aa9cf5fd383633304ddad810
    12592 stmf_sbd: panic in _init on sparc
commit 9d9483ac38cb12576b0b6e2a6d950e6d65a622a0
    12547 pci_pci: NULL pointer errors
commit 692bcae7eb9419bcc3f9d011a67c6a5c6497a72f
    12542 dtrace: NULL pointer errors
commit 12551037071c8ef2216bb540edd94a5bff5e90ae
    12532 unix: NULL pointer errors
commit 13c31a0ee516ab913296f7367d3c09a79ff3ba21
    12533 cfgadm_plugins: NULL pointer errors
commit 67dca347c78f6433be936f7be1c9bbf7069a19ec
    12534 fm: NULL pointer errors

+++ usr/src/uts/common/io/comstar/lu/stmf_sbd/sbd.c
@@ -56,9 +56,9 @@
 extern sbd_status_t sbd_pgr_meta_init(sbd_lu_t *sl);
 extern sbd_status_t sbd_pgr_meta_load(sbd_lu_t *sl);
 extern void sbd_pgr_reset(sbd_lu_t *sl);
-extern int HardwareAcceleratedLocking;
-extern int HardwareAcceleratedInit;
-extern int HardwareAcceleratedMove;
+extern uint8_t HardwareAcceleratedLocking;
+extern uint8_t HardwareAcceleratedInit;
+extern uint8_t HardwareAcceleratedMove;
 extern uint8_t sbd_unmap_enable;
 
 static int sbd_getinfo(dev_info_t *dip, ddi_info_cmd_t cmd, void *arg,
